all things for abstract token now are in an own directory
