+++ cantok/__init__.py
@@ -1,4 +1,4 @@
-from cantok.tokens.abstract_token import AbstractToken as AbstractToken  # noqa: F401
+from cantok.tokens.abstract.abstract_token import AbstractToken as AbstractToken  # noqa: F401
 from cantok.tokens.simple_token import SimpleToken as SimpleToken  # noqa: F401
 from cantok.tokens.condition_token import ConditionToken as ConditionToken  # noqa: F401
 from cantok.tokens.counter_token import CounterToken as CounterToken  # noqa: F401
