+++ cantok/tokens/abstract/cancel_cause.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class CancelCause(Enum):
+    CANCELLED = 1
+    SUPERPOWER = 2
+    NOT_CANCELLED = 3
