+++ tests/units/tokens/test_simple_token.py
@@ -1,6 +1,6 @@
 import pytest
 
-from cantok.tokens.abstract_token import CancelCause, CancellationReport
+from cantok.tokens.abstract.abstract_token import CancelCause, CancellationReport
 from cantok import SimpleToken, CancellationError
